+++ src/components/fields/ComboBox/index.ts
@@ -1,1 +1,2 @@
-export * from './ComboBox';
+export { ComboBox } from './ComboBox';
+export type { CubeComboBoxProps } from './ComboBox';